Move inner classes to top level
